+++ src/threads.js
@@ -65,7 +65,7 @@ StagePickerMorph, CustomBlockDefinition*/
 
 /*jshint esversion: 11, bitwise: false, evil: true*/
 
-modules.threads = '2023-March-10';
+modules.threads = '2023-March-16';
 
 var ThreadManager;
 var Process;
@@ -3356,6 +3356,9 @@ Process.prototype.doPlaySoundUntilDone = function (name) {
     if (name === null || this.context.activeAudio.ended
             || this.context.activeAudio.terminated) {
         if (this.context.activeAudio && this.context.activeAudio.remove) {
+            this.context.activeAudio.currentSrc = null;
+            this.context.activeAudio.src = "";
+            this.context.activeAudio.srcObject = null;
             this.context.activeAudio.remove();
         }
         return null;
@@ -3371,6 +3374,9 @@ Process.prototype.doStopAllSounds = function () {
             if (thread.context) {
                 thread.context.stopMusic();
                 if (thread.context.activeAudio) {
+                    thread.context.activeAudio.currentSrc = null;
+                    thread.context.activeAudio.src = "";
+                    thread.context.activeAudio.srcObject = null;
                     thread.context.activeAudio.remove();
                     thread.popContext();
                 }